Major Update:
Clean Ups:
Many duplicated files were removed.
unnecessary code in files was removed.

UI Fixes:
The text box in the Seach bar is now alligned properly.
the dollar and pound sterling signs were replaced with credits.

Functionality Fixes:
brought back an older working version of the the chat system.
fixed the bug with the reports not being able to be dimessed.